+++ test/libsolidity/util/BytesUtils.cpp
@@ -264,13 +264,13 @@ string BytesUtils::formatBytes(
 			{
 				auto entropy = [](std::string const& str) -> double {
 					double result = 0;
-					map<char, int> frequencies;
+					map<char, double> frequencies;
 					for (char c: str)
 						frequencies[c]++;
 					for (auto p: frequencies)
 					{
-						double freq = static_cast<double>(p.second) / str.length();
-						result -= freq * (log(freq) / log(2));
+						double freq = p.second / double(str.length());
+						result -= freq * (log(freq) / log(2.0));
 					}
 					return result;
 				};
